implemented inversify

+++ Onion arquitecture practice/src/database/data-access.ts
@@ -1,7 +1,9 @@
 import User from './dbUser'
 import IUserRepository from '../repository/user-repository'
 import { UserDataSource } from './data-source'
+import { injectable } from 'inversify'
 
+@injectable()
 export default class UserData implements IUserRepository {
   async create (user: User) {
     const repository = UserDataSource.getRepository(User)
@@ -20,6 +22,12 @@ export default class UserData implements IUserRepository {
 
   async delete (id: number) {
     const repository = UserDataSource.getRepository(User)
-    return await repository.delete({ id })
+    const userById = await repository.findOneBy({ id })
+    if (userById) {
+      await repository.delete({ id })
+      return `User with id ${id} was deleted`
+    } else {
+      return `User with id ${id} wasn't found `
+    }
   }
 }
